+++ mini-spring/src/main/java/initanddestroymethod/beans/factory/DisposableBean.java
@@ -1,0 +1,12 @@
+package initanddestroymethod.beans.factory;
+
+
+import initanddestroymethod.beans.BeansException;
+
+import java.lang.reflect.InvocationTargetException;
+
+public interface DisposableBean {
+
+    void destroy() throws BeansException, InvocationTargetException, IllegalAccessException;
+
+}