+++ mini-spring/src/main/java/initanddestroymethod/beans/factory/HierarchicalBeanFactory.java
@@ -1,0 +1,4 @@
+package initanddestroymethod.beans.factory;
+
+public interface HierarchicalBeanFactory extends BeanFactory {
+}