+++ mini-spring/src/main/java/initanddestroymethod/beans/factory/InitializingBean.java
@@ -1,0 +1,9 @@
+package initanddestroymethod.beans.factory;
+
+import initanddestroymethod.beans.BeansException;
+
+public interface InitializingBean {
+
+    void afterPropertiesSet() throws BeansException;
+
+}