+++ mini-spring/src/main/java/initanddestroymethod/beans/factory/support/BeanDefinitionReader.java
@@ -1,0 +1,24 @@
+package initanddestroymethod.beans.factory.support;
+
+import initanddestroymethod.beans.BeansException;
+import initanddestroymethod.core.io.Resource;
+import initanddestroymethod.core.io.ResourceLoader;
+
+import java.io.IOException;
+
+/**
+ * 读取bean信息即BeanDefinition的接口
+ */
+public interface BeanDefinitionReader {
+
+    BeanDefinitionRegistry getRegistry();
+
+    ResourceLoader getResourceLoader();
+
+    void loadBeanDefinitions(Resource resource) throws IOException;
+
+    void loadBeanDefinitions(String location) throws BeansException;
+
+    void loadBeanDefinitions(String[] locations) throws IOException;
+
+}