+++ mini-spring/src/main/java/initanddestroymethod/beans/factory/support/InstantiationStrategy.java
@@ -1,0 +1,10 @@
+package initanddestroymethod.beans.factory.support;
+
+import initanddestroymethod.beans.BeansException;
+import initanddestroymethod.beans.factory.config.BeanDefinition;
+
+public interface InstantiationStrategy {
+
+    Object instantiate(BeanDefinition beanDefinition) throws BeansException;
+
+}