+++ mini-spring/src/main/java/initanddestroymethod/context/ApplicationContext.java
@@ -1,0 +1,8 @@
+package initanddestroymethod.context;
+
+import initanddestroymethod.beans.factory.HierarchicalBeanFactory;
+import initanddestroymethod.beans.factory.ListableBeanFactory;
+import initanddestroymethod.core.io.ResourceLoader;
+
+public interface ApplicationContext extends ListableBeanFactory, HierarchicalBeanFactory, ResourceLoader {
+}