+++ mini-spring/src/main/java/initanddestroymethod/core/io/ResourceLoader.java
@@ -1,0 +1,13 @@
+package initanddestroymethod.core.io;
+
+import initanddestroymethod.beans.BeansException;
+
+
+/**
+ * 资源加载器接口
+ */
+public interface ResourceLoader {
+
+    Resource getResource(String location) throws BeansException;
+
+}